feat: DiActivity 정의 및 ActivityLifecycleObserver 제거
