+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -3,25 +3,20 @@ package woowacourse.shopping.ui.cart
 import android.content.Context
 import android.os.Bundle
 import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import woowacourse.shopping.R
-import woowacourse.shopping.ShoppingApplication
 import woowacourse.shopping.annotation.Inject
 import woowacourse.shopping.annotation.ReleasedActivityLifecycle
 import woowacourse.shopping.annotation.RetainedActivityLifecycle
 import woowacourse.shopping.databinding.ActivityCartBinding
 import woowacourse.shopping.di.ActivityClassName
-import woowacourse.shopping.lifecycleobserver.ActivityLifecycleObserver
-import woowacourse.shopping.lifecycleobserver.DefaultActivityLifecycleObserver
+import woowacourse.shopping.ui.DiAppCompatActivity
 import woowacourse.shopping.ui.viewModelFactory
 
-fun createDateFormatter(@ReleasedActivityLifecycle(ActivityClassName.CART_ACTIVITY) context: Context): DateFormatter = DateFormatter(context)
+fun createDateFormatter(@ReleasedActivityLifecycle(ActivityClassName.CART_ACTIVITY) context: Context): DateFormatter =
+    DateFormatter(context)
 
-class CartActivity :
-    AppCompatActivity(),
-    ActivityLifecycleObserver by DefaultActivityLifecycleObserver() {
-
+class CartActivity : DiAppCompatActivity() {
     private val binding by lazy { ActivityCartBinding.inflate(layoutInflater) }
 
     private val viewModel by lazy {
@@ -38,8 +33,6 @@ class CartActivity :
         setupBinding()
         setupToolbar()
         setupView()
-
-        setupLifecycleObserver()
     }
 
     override fun onSupportNavigateUp(): Boolean {
@@ -80,14 +73,5 @@ class CartActivity :
             if (!it) return@observe
             Toast.makeText(this, getString(R.string.cart_deleted), Toast.LENGTH_SHORT).show()
         }
-    }
-
-    private fun setupLifecycleObserver() {
-        val shoppingApplication = application as ShoppingApplication
-        setupLifecycleObserver(
-            lifecycle = lifecycle,
-            activity = this,
-            injector = shoppingApplication.injector,
-        )
     }
 }
